# project-management-backend/src/main/java/com/biswas/project_management_backend/controller/TaskController.java
package com.biswas.project_management_backend.controller;

import com.biswas.project_management_backend.dto.TaskDto;
import com.biswas.project_management_backend.dto.mapper.TaskDtoMapper;
import com.biswas.project_management_backend.service.TaskService;
import lombok.RequiredArgsConstructor;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;

import java.util.List;

@RestController
@RequestMapping("/api/tasks")
@RequiredArgsConstructor
public class TaskController {

    private final TaskService taskService;

    // Get all tasks
    @GetMapping("/company/{companyId}")
    public ResponseEntity<List<TaskDto>> getAllTasks(@PathVariable Long companyId) {
        List<TaskDto> tasks = taskService.getAllTasks(companyId);

        return ResponseEntity.ok(tasks);
    }

    // Get a single task by ID
    @GetMapping("/{id}")
    public ResponseEntity<TaskDto> getTaskById(@PathVariable Long id) {
        TaskDto task = taskService.getTaskById(id);
        return ResponseEntity.ok(task);
    }

    @PostMapping
    public ResponseEntity<TaskDto> createTask(@RequestBody TaskDto dto) {
        TaskDto created = taskService.createTask(dto);
        return ResponseEntity.ok(created);
    }

    @PutMapping("/{id}")
    public ResponseEntity<TaskDto> updateTask(@PathVariable Long id, @RequestBody TaskDto dto) {
        TaskDto updated = taskService.updateTask(id, dto);
        return ResponseEntity.ok(updated);
    }

    @DeleteMapping("/{id}")
    public ResponseEntity<Void> deleteTask(@PathVariable Long id) {
        taskService.deleteTask(id);
        return ResponseEntity.noContent().build();
    }

    @GetMapping("/project/{projectId}")
    public ResponseEntity<List<TaskDto>> getTasksByProject(@PathVariable Long projectId) {
        List<TaskDto> tasks = taskService.getTasksByProject(projectId);

        return ResponseEntity.ok(tasks);
    }
}